+++ R/ssa_methods.R
@@ -11,18 +11,18 @@ ssa_method <- function(name, params, factory) {
 
 #' Euler-Maruyama method (EM)
 #'
-#' Euler-Maruyama method implementatio of the \acronym{SDE} as described by Euler and Maruyama (?).
+#' Euler-Maruyama method implementation of the \acronym{ODE}.
 #'
 #' @param tau tau parameter
 #' @param noise_strength noise_strength parameter
 #'
 #' @export
-ssa_em <- function(tau = 0.01, noise_strength = 2) {
+ode_em <- function(tau = 0.01, noise_strength = 2) {
   ssa_method(
     name = "EM",
     params = lst(tau, noise_strength),
     factory = function() {
-      make_ssa_em(tau, noise_strength)
+      make_ode_em(tau, noise_strength)
     }
   )
 }