+++ src/RcppExports.cpp
@@ -5,6 +5,18 @@
 
 using namespace Rcpp;
 
+// make_ode_em
+SEXP make_ode_em(double tau, double noise_strength);
+RcppExport SEXP _fastgssa_make_ode_em(SEXP tauSEXP, SEXP noise_strengthSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
+    Rcpp::traits::input_parameter< double >::type noise_strength(noise_strengthSEXP);
+    rcpp_result_gen = Rcpp::wrap(make_ode_em(tau, noise_strength));
+    return rcpp_result_gen;
+END_RCPP
+}
 // simulate
 List simulate(SEXP propensity_funs, const int num_functions, SEXP ssa_alg, const NumericVector& initial_state, const NumericVector& params, const IntegerVector& nu_i, const IntegerVector& nu_p, const IntegerVector& nu_x, const double final_time, const double census_interval, const int buffer_size, const double max_walltime, const bool stop_on_neg_state, const bool verbose, const double console_interval);
 RcppExport SEXP _fastgssa_simulate(SEXP propensity_funsSEXP, SEXP num_functionsSEXP, SEXP ssa_algSEXP, SEXP initial_stateSEXP, SEXP paramsSEXP, SEXP nu_iSEXP, SEXP nu_pSEXP, SEXP nu_xSEXP, SEXP final_timeSEXP, SEXP census_intervalSEXP, SEXP buffer_sizeSEXP, SEXP max_walltimeSEXP, SEXP stop_on_neg_stateSEXP, SEXP verboseSEXP, SEXP console_intervalSEXP) {
@@ -51,18 +63,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// make_ssa_em
-SEXP make_ssa_em(double tau, double noise_strength);
-RcppExport SEXP _fastgssa_make_ssa_em(SEXP tauSEXP, SEXP noise_strengthSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< double >::type tau(tauSEXP);
-    Rcpp::traits::input_parameter< double >::type noise_strength(noise_strengthSEXP);
-    rcpp_result_gen = Rcpp::wrap(make_ssa_em(tau, noise_strength));
-    return rcpp_result_gen;
-END_RCPP
-}
 // make_ssa_etl
 SEXP make_ssa_etl(double tau);
 RcppExport SEXP _fastgssa_make_ssa_etl(SEXP tauSEXP) {
@@ -76,10 +76,10 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_fastgssa_make_ode_em", (DL_FUNC) &_fastgssa_make_ode_em, 2},
     {"_fastgssa_simulate", (DL_FUNC) &_fastgssa_simulate, 15},
     {"_fastgssa_make_ssa_btl", (DL_FUNC) &_fastgssa_make_ssa_btl, 1},
     {"_fastgssa_make_ssa_direct", (DL_FUNC) &_fastgssa_make_ssa_direct, 0},
-    {"_fastgssa_make_ssa_em", (DL_FUNC) &_fastgssa_make_ssa_em, 2},
     {"_fastgssa_make_ssa_etl", (DL_FUNC) &_fastgssa_make_ssa_etl, 1},
     {NULL, NULL, 0}
 };