+++ src/ode_em.cpp
@@ -3,9 +3,9 @@
 
 using namespace Rcpp;
 
-class SSA_EM : public SSA {
+class ODE_EM : public SSA {
 public:
-   SSA_EM(double tau_ = .01, double noise_strength_ = 2.0) : SSA("EM"), tau(tau_), noise_strength(noise_strength_) {}
+  ODE_EM(double tau_ = .01, double noise_strength_ = 2.0) : SSA("EM"), tau(tau_), noise_strength(noise_strength_) {}
 
   double tau;
   double noise_strength;
@@ -36,8 +36,8 @@ public:
 } ;
 
 // [[Rcpp::export]]
-SEXP make_ssa_em(double tau, double noise_strength) {
-  SSA_EM *ssa = new SSA_EM(tau, noise_strength);
-  XPtr<SSA_EM> ptr(ssa);
+SEXP make_ode_em(double tau, double noise_strength) {
+  ODE_EM *ssa = new ODE_EM(tau, noise_strength);
+  XPtr<ODE_EM> ptr(ssa);
   return ptr;
 }